Split out records module
